moving maploading code to src
