+++ Cargo.toml
@@ -7,6 +7,8 @@ edition = "2024"
 [dependencies]
 bevy.workspace = true
 rand.workspace = true
+serde = { version = "1.0", features = ["derive"] }
+serde_json = "1.0"
 
 [profile.dev]
 opt-level = 1
@@ -38,12 +40,8 @@ rand = "0.9.2"
 
 [workspace]
 members = [
-    "slideshow",
-    "ai",
-    "movementsystem",
-    "multiplayer",
-    "maploading",
     "prototype",
+    "maploading",
+    "slideshow",
     "xtask",
-    "platform",
 ]
