+++ src/main.rs
@@ -1,3 +1,5 @@
+mod util;
+mod map;
 mod app;
 mod player;
 mod config;
