+++ src/map/loader.rs
