+++ src/physics/integrate.rs
@@ -49,18 +49,18 @@ pub fn boundary(
     let width = 1280.0 - 64.0; // minus player width
     let height = 720.0 - 64.0; // minus player height
     for (mut transform, mut velocity, mut momentum) in query.iter_mut() {
-        if transform.translation.x < - width / 2.0 {
-            transform.translation.x = - width / 2.0;
+        if transform.translation.x < 32.0{
+            transform.translation.x = 32.0;
             velocity.0.x = 0.0;
             momentum.0.x = 0.0;
         }
-        if transform.translation.x > width / 2.0 {
-            transform.translation.x = width / 2.0;
+        if transform.translation.x > width + 32.0{
+            transform.translation.x = width + 32.0;
             velocity.0.x = 0.0;
             momentum.0.x = 0.0;
         }
-        if transform.translation.y < - height / 2.0 {
-            transform.translation.y = - height / 2.0;
+        if transform.translation.y < 32.0 {
+            transform.translation.y = 32.0;
             velocity.0.y = 0.0;
             momentum.0.y = 0.0;
         }
