+++ src/util/mod.rs
@@ -1,2 +1,1 @@
-pub mod map;
 pub mod dev_mode;
