--- workspaces/maploading/Cargo.toml
@@ -1,15 +1,0 @@
-[package]
-name = "maploading"
-
-edition.workspace = true
-version.workspace = true
-authors.workspace = true
-
-[dependencies]
-bevy.workspace = true
-rand.workspace = true
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0"
-base64 = "0.22"
-bevy_ecs_tilemap = "0.16.0"
-bevy_image = "0.16.1"
