--- workspaces/maploading/README.md
@@ -1,6 +1,0 @@
-# README
-if you wish to run anything please use 
-
-cargo run -p {foldername}
-
-or if you are in a directory use the cargo run
