--- workspaces/maploading/main.rs
@@ -1,9 +1,0 @@
-mod app;
-mod player;
-mod config;
-mod components;
-mod physics;
-
-fn main() {
-    app::run();
-}
