--- workspaces/maploading/src/components/mod.rs
@@ -1,17 +1,0 @@
-// SPDX-License-Identifier: MIT
-// Copyright (c) 2025 Tingxu Chen
-// Author: Tingxu Chen <tic128@pitt.edu>
-// Description: <Player bundle and components>
-use bevy::prelude::*;
-
-pub mod motion;
-pub mod rope;
-pub mod collision;
-
-pub mod prelude {
-    pub use super::motion::Velocity;
-    pub use super::motion::NetForce;
-    pub use super::motion::ControlForce;
-    pub use super::rope::Rope;
-    pub use super::collision::Aabb;
-}
