--- workspaces/maploading/src/config/mod.rs
@@ -1,10 +1,0 @@
-// SPDX-License-Identifier: MIT
-// Copyright (c) 2025 Tingxu Chen
-// Author: Tingxu Chen <tic128@pitt.edu>
-// Description: <Configuration module>
-
-pub mod player;
-pub mod physics;
-
-pub use player::*;
-pub use physics::*;
