--- workspaces/maploading/src/main.rs
@@ -1,10 +1,0 @@
-mod util;
-mod app;
-mod player;
-mod config;
-mod components;
-mod physics;
-
-fn main() {
-    app::run();
-}
