--- workspaces/maploading/src/player/mod.rs
@@ -1,5 +1,0 @@
-pub mod bundle;
-pub mod player_plugin;
-
-pub use player_plugin::PlayerPlugin;
-pub use bundle::{Player, PlayerControls, PlayerBundle};
